Update manager Test Code 작성 Redis를 사용하는 기능은 제외 한 나머지 기능 구현 완료 Service DTO Test Code

+++ src/manages/manages.service.ts
@@ -15,7 +15,6 @@ export class ManagesService {
   //스테이지 정보 확인
   async findStageInfo(param: GameInfoDTO) {
     const { stage } = param;
-
     return await this.managesRepository.findStageInfo(stage);
   }
 
@@ -61,6 +60,11 @@ export class ManagesService {
   //유저 삭제
   async deleteUser(param: UserSelectDTO) {
     const { id } = param;
+    const userCheck = await this.managesRepository.findUserStatus(id);
+
+    if (!userCheck) {
+      throw new NotFoundException();
+    }
 
     await this.managesRepository.deleteUser(id);
     return '';
